+++ clients/logs/ecs_cloudwatch_logs_client_test.go
